move all event reg option components up 1 dir
